services/libs/pyproject.toml
[project]
name = "huleedu-service-libs"
version = "0.2.0"
description = "HuleEdu Shared Service Libraries"
authors = [
    { name = "Olof Larsson", email = "paunchygent@gmail.com" },
]
requires-python = ">=3.11"
dependencies = [
    "aiokafka",
    "pydantic",
    "structlog",
    "huleedu-common-core",
    "redis[hiredis]>=5.0.0",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = false

[tool.mypy]
python_version = "3.11"
warn_return_any = true
ignore_missing_imports = false
check_untyped_defs = true
disallow_untyped_defs = false

# External libraries without official type stubs
[[tool.mypy.overrides]]
module = [
    "aiokafka.*",
    "redis.*",
]
ignore_missing_imports = true 